+++ spec/legion/extensions/node/transport/queues/cluster_control_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
+require 'legion/extensions/node/config'
 
 # Mirror ClusterControl queue logic in a proxy class to avoid superclass mismatch
 # with Legion::Transport::Queue which is not available in the test environment.
@@ -10,13 +11,36 @@ CLUSTER_CONTROL_QUEUE_TEST_CLASS = Class.new do
   end
 
   def queue_options
-    { durable: true, exclusive: false, auto_delete: false,
-      arguments: {
-        'x-queue-type':  'classic',
-        'x-expires':     604_800_000,
-        'x-message-ttl': 86_400_000,
-        'x-max-length':  1000
-      } }
+    options = Legion::Extensions::Node::Config.control_queue
+    { durable:     truthy?(options[:durable], default: true),
+      exclusive:   truthy?(options[:exclusive], default: false),
+      auto_delete: truthy?(options[:auto_delete], default: false),
+      arguments:   queue_arguments(options) }
+  end
+
+  private
+
+  def queue_arguments(options)
+    {
+      'x-queue-type':  options[:queue_type],
+      'x-expires':     positive_integer(options[:expires_ms]),
+      'x-message-ttl': positive_integer(options[:message_ttl_ms]),
+      'x-max-length':  positive_integer(options[:max_length])
+    }.compact
+  end
+
+  def positive_integer(value)
+    return value if value.is_a?(Integer) && value.positive?
+    return nil unless value.to_s.match?(/\A\d+\z/)
+
+    parsed = value.to_i
+    parsed.positive? ? parsed : nil
+  end
+
+  def truthy?(value, default:)
+    return default if value.nil?
+
+    value == true || value.to_s == 'true'
   end
 end
 
@@ -24,6 +48,8 @@ RSpec.describe 'ClusterControl queue' do
   subject(:queue) { CLUSTER_CONTROL_QUEUE_TEST_CLASS.new }
 
   before do
+    allow(Legion::Settings).to receive(:dig).and_return(nil)
+    allow(Legion::Settings).to receive(:dig).with(:extensions, :node).and_return({})
     allow(Legion::Settings).to receive(:[]).with(:client).and_return({ name: 'test-node' })
   end
 
@@ -55,6 +81,31 @@ RSpec.describe 'ClusterControl queue' do
         'x-expires':     604_800_000,
         'x-message-ttl': 86_400_000,
         'x-max-length':  1000
+      )
+    end
+
+    it 'uses configured queue durability and retention overrides' do
+      allow(Legion::Settings).to receive(:dig).with(:extensions, :node).and_return(
+        cluster_control: {
+          queue: {
+            durable:        false,
+            auto_delete:    true,
+            expires_ms:     nil,
+            message_ttl_ms: 30_000,
+            max_length:     25
+          }
+        }
+      )
+
+      expect(queue.queue_options).to eq(
+        durable:     false,
+        exclusive:   false,
+        auto_delete: true,
+        arguments:   {
+          'x-queue-type':  'classic',
+          'x-message-ttl': 30_000,
+          'x-max-length':  25
+        }
       )
     end
   end